requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.6
Flask-MySQL==1.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==4.25.3
PyMySQL==1.1.0
PyYAML==6.0.1
SQLAlchemy==1.4.30
typing_extensions==4.10.0
Werkzeug==3.0.1
